AST-804 Make sure test are deterministic, fix precision in correction sums

+++ main/wsclean.cpp
@@ -1224,8 +1224,9 @@ void WSClean::partitionSingleGroup(const ImagingTable& facetGroup,
     if (!isPredictOnly) {
       if (_settings.applyFacetBeam || !_settings.facetSolutionFiles.empty()) {
         // Apply average direction dependent correction before storing
-        const float m = _msGridderMetaCache[facetEntry.index]->correctionSum /
-                        facetEntry.imageWeight;
+        const long double m =
+            _msGridderMetaCache[facetEntry.index]->correctionSum /
+            facetEntry.imageWeight;
         facetImage *= 1.0f / std::sqrt(m);
       }
     }
@@ -1811,8 +1812,9 @@ void WSClean::stitchSingleGroup(const ImagingTable& facetGroup,
 
     if (!isPSF &&
         (_settings.applyFacetBeam || !_settings.facetSolutionFiles.empty())) {
-      const float m = _msGridderMetaCache[facetEntry.index]->correctionSum /
-                      facetEntry.imageWeight;
+      const long double m =
+          _msGridderMetaCache[facetEntry.index]->correctionSum /
+          facetEntry.imageWeight;
       facetImage *= 1.0f / std::sqrt(m);
     }
 
@@ -2155,8 +2157,8 @@ void WSClean::correctImagesH5(aocommon::FitsWriter& writer,
       facetImage.SetFacet(*entry.facet, true);
       // Requires std::map::at in order to comply with constness of member
       // function
-      const float m = _msGridderMetaCache.at(entry.index)->correctionSum /
-                      entry.imageWeight;
+      const long double m = _msGridderMetaCache.at(entry.index)->correctionSum /
+                            entry.imageWeight;
       facetImage.MultiplyImageInsideFacet(imagePtr, 1.0f / std::sqrt(m));
     }
