+++ scheduling/metadatacache.h
@@ -12,9 +12,16 @@ struct MetaDataCache {
         integrationTime;
   };
   std::vector<Entry> msDataVector;
-  float h5Sum = 0.0;
-  float correctionSum = 0.0;
-
+  /** @{
+   *  These variables are incremented with a comparatively small value for each
+   * gridded visibility, hence a long double is used to accomodate sufficient
+   * precision.
+   */
+  long double h5Sum = 0.0;
+  long double correctionSum = 0.0;
+  /**
+   * @}
+   */
   void Serialize(aocommon::SerialOStream& stream) const;
   void Unserialize(aocommon::SerialIStream& stream);
 };
